backend/requirements.txt
flask==2.3.2
flask-socketio==5.3.4
flask-sqlalchemy==3.0.3
flask-cors==4.0.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
eventlet==0.33.3
